Hello, I have started to use your library to detect anomalies in our system health metrics. I am still learning Go but enjoy working the Go benchmarks. I have left your existing public methods along and have added two new methods. A `ClearDataSet` which will nil out the `dataSet      []big.Float` allowing for that information to garbage collected. Secondly `EventXIsAnomalous(eventX, threshold *big.Float) (bool, *big.Float)`, this is a mutation of the existing `EventIsAnomalous` which saves one memory allocation. I did not want to break any existing code by altering the method signature. 

One downside is that the `EventIsAnomalous` is harder to read because I have collapsed the operations into as few steps and have reused a variable `tempA` to save on allocations.

Now for the numbers. The old is from commit 5ba9cc3731c04205783cedd103595ba2db4346b3 and new is from commit b652c5ef5edb3c76f5d995d00f3f1a9b28f71a4b
```
benchmark                           old ns/op     new ns/op     delta
BenchmarkNewAnomalyDetection-8      10030         9839          -1.90%
BenchmarkExpandDataSet-8            21531         10454         -51.45%
BenchmarkEventIsAnomalousn-8        1302          807           -38.02%
BenchmarkEventXIsAnomalous-8        1285          744           -42.10%
BenchmarkCalculateProbability-8     1251          727           -41.89%
BenchmarkEstimateVariance-8         15055         14957         -0.65%

benchmark                           old allocs     new allocs     delta
BenchmarkNewAnomalyDetection-8      75             68             -9.33%
BenchmarkExpandDataSet-8            134            67             -50.00%
BenchmarkEventIsAnomalousn-8        22             9              -59.09%
BenchmarkEventXIsAnomalous-8        22             8              -63.64%
BenchmarkCalculateProbability-8     21             8              -61.90%
BenchmarkEstimateVariance-8         86             83             -3.49%

benchmark                           old bytes     new bytes     delta
BenchmarkNewAnomalyDetection-8      3680          3344          -9.13%
BenchmarkExpandDataSet-8            8080          3120          -61.39%
BenchmarkEventIsAnomalousn-8        992           368           -62.90%
BenchmarkEventXIsAnomalous-8        992           320           -67.74%
BenchmarkCalculateProbability-8     944           320           -66.10%
BenchmarkEstimateVariance-8         4064          3920          -3.54%
```
 
Let me know if you have any comments or concerns.

If you could share a guide to the math that would be amazing. I would like to be able to run a rolling mean and variance so that the `dataSet      []big.Float` does not need to be stored in the AnomalyDetection and could be expanded at any time.