@silenteh 2017-05-22
Thanks for this pull request!
The math was taken from Andrew NG Coursera class on machine learning.